+++ locales/en/translation.json
@@ -1,9 +1,12 @@
 {
-  "Settings": "■ Settings",
+  "Description": "■ About App",
+  "BasedSettings": "■ Basic Settings",
   "AISettings": "■ AI Settings",
+  "CharacterSettings": "■ Character Settings",
   "YoutubeSettings": "■ YouTube Settings",
   "VoiceSettings": "■ Voice Settings",
   "SlideSettings": "■ Slide Settings",
+  "LogSettings": "■ Conversation History",
   "OtherSettings": "■ Other",
   "ExternalLinkageMode": "External Linkage Mode (WebSocket)",
   "YoutubeMode": "YouTube Mode",
@@ -48,8 +51,8 @@
   "OpenVRM": "Open VRM",
   "BackgroundImage": "Background Image",
   "ChangeBackgroundImage": "Change Background Image",
-  "CharacterSettingsPrompt": "Character Settings (System Prompt)",
-  "CharacterSettingsInfo": "Character settings are set in the system prompt. By specifying emotion tags, you can control the character's expression and motion. Example: [neutral]Good morning![happy]Good luck today!",
+  "CharacterSettingsPrompt": "Character Prompt",
+  "CharacterSettingsInfo": "This value is set as the system prompt.\nPlease refer to the initial prompt and specify the emotion tags to control the character's expressions and motions. Example: [neutral]Good morning![happy]Today is also a hard day!",
   "CharacterSettingsReset": "Reset Character Settings",
   "SyntheticVoiceEngineChoice": "Choose Synthetic Voice Engine",
   "VoiceAdjustment": "Voice Adjustment",
@@ -103,6 +106,7 @@
   "AnswerGenerating": "Answer Generating",
   "AboutThisApplication": "About This Application",
   "AboutThisApplicationDescription": "Enjoy conversations with a 3D character right in your web browser, using microphone or text input and voice synthesis. You can also change the character (VRM), adjust its personality, and modify its voice.<br />Settings can be changed from the menu button in the top left.",
+  "AboutThisApplicationDescription2": "If you want to change the character, please refer to the \"Character Settings\" tab.",
   "TechnologyIntroduction": "Technology Introduction",
   "TechnologyIntroductionDescription1": "This app was created by modifying pixiv's <b>ChatVRM</b>. The original source code can be found",
   "TechnologyIntroductionLink1": "here",
@@ -118,6 +122,10 @@
   "RepositoryURL": "Repository URL:",
   "DontShowIntroductionNextTime": "Do not show this dialog next time",
   "Close": "CLOSE",
+  "Contact": "Contact",
+  "ContactDescription": "Please contact me via the email address or Twitter account below regarding this app.",
+  "Creator": "Creator",
+  "CreatorDescription": "Creator: Tegan",
   "Language": "Language",
   "UsingGSVITTS": "GSVI TTS",
   "GSVITTSInfo": "GSVI TTS Settings",
