+++ locales/ja/translation.json
@@ -1,9 +1,12 @@
 {
-  "Settings": "■ 基本設定",
+  "Description": "■ アプリについて",
+  "BasedSettings": "■ 基本設定",
   "AISettings": "■ AI設定",
+  "CharacterSettings": "■ キャラクター設定",
   "YoutubeSettings": "■ YouTube設定",
   "VoiceSettings": "■ 音声設定",
   "SlideSettings": "■ スライド設定",
+  "LogSettings": "■ 会話履歴",
   "OtherSettings": "■ その他",
   "ExternalLinkageMode": "外部連携モード（WebSocket）",
   "YoutubeMode": "YouTubeモード",
@@ -48,8 +51,8 @@
   "OpenVRM": "VRMを開く",
   "BackgroundImage": "背景画像",
   "ChangeBackgroundImage": "背景画像を変える",
-  "CharacterSettingsPrompt": "キャラクター設定（システムプロンプト）",
-  "CharacterSettingsInfo": "キャラクター設定はシステムプロンプトに設定されます。\n初期プロンプトを参考に、感情タグを指定することでキャラクターの表情やモーションを制御できます。例: [neutral]おはようございます！[happy]今日もお疲れ様です！",
+  "CharacterSettingsPrompt": "キャラクタープロンプト",
+  "CharacterSettingsInfo": "この値はシステムプロンプトとして設定されます。\n初期プロンプトを参考に、感情タグを指定することでキャラクターの表情やモーションを制御できます。例: [neutral]おはようございます！[happy]今日もお疲れ様です！",
   "CharacterSettingsReset": "キャラクター設定リセット",
   "SyntheticVoiceEngineChoice": "合成音声エンジンの選択",
   "VoiceAdjustment": "声の調整",
@@ -104,6 +107,7 @@
   "AnswerGenerating": "回答生成中",
   "AboutThisApplication": "このアプリケーションについて",
   "AboutThisApplicationDescription": "Webブラウザだけで3Dキャラクターとの会話を、マイクやテキスト入力、音声合成を用いて楽しめます。キャラクター（VRM）の変更や性格設定、音声調整もできます。<br />設定は左上のメニューボタンから変更できます。",
+  "AboutThisApplicationDescription2": "AITuberKitでは、WebブラウザだけでAIキャラクターとの会話を楽しめます。キャラクターの変更や性格設定、音声調整は各設定項目を確認してください。",
   "TechnologyIntroduction": "技術紹介",
   "TechnologyIntroductionDescription1": "このアプリはpixiv社の<b>ChatVRM</b>を改造して作成されています。元のソースコードは",
   "TechnologyIntroductionLink1": "こちら",
@@ -119,6 +123,10 @@
   "RepositoryURL": "リポジトリURL:",
   "DontShowIntroductionNextTime": "次回からこのダイアログを表示しない",
   "Close": "閉じる",
+  "Contact": "お問い合わせ",
+  "ContactDescription": "このアプリに関するお問い合わせは下記のメールアドレスまたはTwitterアカウントにお願いします。",
+  "Creator": "作成者情報",
+  "CreatorDescription": "作成者: ニケ",
   "Language": "言語設定",
   "UsingGSVITTS": "GSVI TTSを使用する",
   "GSVITTSInfo": "GSVI TTS設定",
