+++ locales/ko/translation.json
@@ -1,9 +1,12 @@
 {
-  "Settings": "■ 설정",
+  "Description": "■ 앱에 대해",
+  "BasedSettings": "■ 기본 설정",
   "AISettings": "■ AI 설정",
+  "CharacterSettings": "■ 캐릭터 설정",
   "YoutubeSettings": "■ YouTube 설정",
   "VoiceSettings": "■ 음성 설정",
   "SlideSettings": "■ 슬라이드 설정",
+  "LogSettings": "■ 대화 기록",
   "OtherSettings": "■ 기타",
   "ExternalLinkageMode": "외부 연동 모드 (WebSocket)",
   "YoutubeMode": "YouTube 모드",
@@ -48,8 +51,8 @@
   "OpenVRM": "VRM 열기",
   "BackgroundImage": "배경 이미지",
   "ChangeBackgroundImage": "배경 이미지 변경",
-  "CharacterSettingsPrompt": "캐릭터 설정 (시스템 프롬프트)",
-  "CharacterSettingsInfo": "캐릭터 설정은 시스템 프롬프트에 설정됩니다. 초기 프롬프트를 참고하여 감정 태그를 지정하여 캐릭터의 표정과 모션을 제어할 수 있습니다. 예: [neutral]안녕하세요![happy]오늘도 힘내세요!",
+  "CharacterSettingsPrompt": "캐릭터 프롬프트",
+  "CharacterSettingsInfo": "이 값은 시스템 프롬프트로 설정됩니다.\n초기 프롬프트를 참고하여 감정 태그를 지정하여 캐릭터의 표정과 모션을 제어할 수 있습니다. 예: [neutral]좋은 아침이에요![happy]오늘도 힘든 하루예요!",
   "CharacterSettingsReset": "캐릭터 설정 재설정",
   "SyntheticVoiceEngineChoice": "합성 음성 엔진 선택",
   "VoiceAdjustment": "목소리 조정",
@@ -103,6 +106,7 @@
   "AnswerGenerating": "답변 생성중",
   "AboutThisApplication": "이 애플리케이션에 대해",
   "AboutThisApplicationDescription": "웹 브라우저에서 3D 캐릭터와 대화를 즐길 수 있습니다. 마이크나 텍스트 입력, 음성 합성을 사용하여 대화할 수 있습니다. 캐릭터(VRM)의 변경이나 성격 설정, 목소리 조정도 가능합니다. 설정은 왼쪽 상단의 메뉴 버튼에서 변경할 수 있습니다.",
+  "AboutThisApplicationDescription2": "캐릭터를 변경하려면 \"캐릭터 설정\" 탭을 참고하세요.",
   "TechnologyIntroduction": "기술 소개",
   "TechnologyIntroductionDescription1": "이 애플리케이션은 pixiv사의<b>ChatVRM</b>을 개조하여 만들어졌습니다. 원본 소스 코드는",
   "TechnologyIntroductionLink1": "여기",
@@ -118,6 +122,10 @@
   "RepositoryURL": "리포지토리 URL:",
   "DontShowIntroductionNextTime": "다음에 이 대화 상자를 표시하지 않기",
   "Close": "닫기",
+  "Contact": "문의",
+  "ContactDescription": "이 애플리케이션에 대한 문의는 아래 이메일 주소 또는 Twitter 계정을 통해 문의하십시오.",
+  "Creator": "제작자 정보",
+  "CreatorDescription": "제작자: 티엔",
   "Language": "언어 설정",
   "UsingGSVITTS": "GSVI TTS 사용",
   "GSVITTSInfo": "GSVI TTS 설정",
