+++ locales/zh/translation.json
@@ -1,9 +1,12 @@
 {
-  "Settings": "■ 設定",
+  "Description": "■ 關於這個應用程式",
+  "BasedSettings": "■ 基本設定",
   "AISettings": "■ AI 設定",
+  "CharacterSettings": "■ 角色設定",
   "YoutubeSettings": "■ YouTube 設定",
   "VoiceSettings": "■ 語音設定",
   "SlideSettings": "■ 投影片設定",
+  "LogSettings": "■ 會話記錄",
   "OtherSettings": "■ 其他",
   "ExternalLinkageMode": "外部連線模式 (WebSocket）",
   "YoutubeMode": "YouTube 模式",
@@ -48,8 +51,8 @@
   "OpenVRM": "開啟 VRM",
   "BackgroundImage": "背景圖片",
   "ChangeBackgroundImage": "更改背景圖片",
-  "CharacterSettingsPrompt": "角色設定 (系統提示)",
-  "CharacterSettingsInfo": "角色設定是設定在系統提示中。\n請參考初始提示，並指定感情標籤，以控制角色的表情和動作。\n例: [neutral]早上好！[happy]今天也要加油喔！",
+  "CharacterSettingsPrompt": "角色提示",
+  "CharacterSettingsInfo": "此值設置為系統提示。\n請參考初始提示，並指定感情標籤，以控制角色的表情和動作。\n例: [neutral]早上好！[happy]今天也要加油喔！",
   "CharacterSettingsReset": "重設角色設定",
   "SyntheticVoiceEngineChoice": "選擇合成語音引擎",
   "VoiceAdjustment": "語音調整",
@@ -103,6 +106,7 @@
   "AnswerGenerating": "正在生成回答",
   "AboutThisApplication": "關於這個應用程式",
   "AboutThisApplicationDescription": "只需使用網頁瀏覽器，即可透過麥克風或文字輸入、語音合成與 3D 角色對話。你可以更換角色（VRM）、設定性格和調整語音。<br />設定可以從左上角的選單按鈕進行更改。",
+  "AboutThisApplicationDescription2": "如果想要更換角色，請參考「角色設定」標籤。",
   "TechnologyIntroduction": "技術介紹",
   "TechnologyIntroductionDescription1": "這個應用程式是基於 pixiv 公司的 <b>ChatVRM</b> 改造而成。原始程式碼請參考",
   "TechnologyIntroductionLink1": "這裡",
@@ -118,6 +122,10 @@
   "RepositoryURL": "倉庫 URL：",
   "DontShowIntroductionNextTime": "下次不再顯示此對話方塊",
   "Close": "關閉",
+  "Contact": "聯絡我們",
+  "ContactDescription": "有關此應用程式的任何問題，請透過以下電子郵件地址或 Twitter 帳號聯絡我們。",
+  "Creator": "創作者資訊",
+  "CreatorDescription": "創作者：Tegan",
   "Language": "語言設定",
   "UsingGSVITTS": "使用 GSVI TTS",
   "GSVITTSInfo": "GSVI TTS 設定",
