+++ src/components/settings/log.tsx
@@ -23,9 +23,9 @@ const Log = () => {
   }
 
   return (
-    <div className="mt-40">
-      <div className="my-8 grid-cols-2">
-        <div className="my-16 typography-20 font-bold">
+    <div className="">
+      <div className="mb-8 grid-cols-2">
+        <div className="mb-16 typography-20 font-bold">
           {t('ConversationHistory')}
         </div>
         <div className="my-8">
