+++ src/components/settings/modelProvider.tsx
@@ -2,7 +2,6 @@ import { useTranslation } from 'react-i18next'
 import menuStore from '@/features/stores/menu'
 import settingsStore from '@/features/stores/settings'
 import slideStore from '@/features/stores/slide'
-import { SYSTEM_PROMPT } from '@/features/constants/systemPromptConstants'
 import { Link } from '../link'
 import { TextButton } from '../textButton'
 import { useCallback } from 'react'
@@ -1086,34 +1085,6 @@ const ModelProvider = () => {
           </div>
         </>
       )}
-      <div className="mt-40">
-        <div className="my-8">
-          <div className="my-16 typography-20 font-bold">
-            {t('CharacterSettingsPrompt')}
-          </div>
-          {selectAIService === 'dify' ? (
-            <div className="my-16">{t('DifyInstruction')}</div>
-          ) : (
-            <div className="my-16 whitespace-pre-line">
-              {t('CharacterSettingsInfo')}
-            </div>
-          )}
-          <TextButton
-            onClick={() =>
-              settingsStore.setState({ systemPrompt: SYSTEM_PROMPT })
-            }
-          >
-            {t('CharacterSettingsReset')}
-          </TextButton>
-        </div>
-        <textarea
-          value={systemPrompt}
-          onChange={(e) =>
-            settingsStore.setState({ systemPrompt: e.target.value })
-          }
-          className="px-16 py-8 bg-surface1 hover:bg-surface1-hover h-168 rounded-8 w-full"
-        ></textarea>
-      </div>
     </div>
   )
 }
