+++ src/components/settings/other.tsx
@@ -1,13 +1,11 @@
 import AdvancedSettings from './advancedSettings'
 import MessageReceiverSetting from './messageReceiver'
-import Log from './log'
 
 const Other = () => {
   return (
     <>
       <AdvancedSettings />
       <MessageReceiverSetting />
-      <Log />
     </>
   )
 }
